problems with CSRF need fix

+++ src/main/java/com/github/simbir_hospital_service/controller/HospitalController.java
@@ -1,92 +1,3 @@
-//package com.github.simbir_hospital_service.controller;
-//
-//import com.github.simbir_hospital_service.client.AccountServiceClient;
-//import com.github.simbir_hospital_service.config.context.UserContext;
-//import com.github.simbir_hospital_service.hospital.request.GetHospitalRequest;
-//import com.github.simbir_hospital_service.hospital.Hospital;
-//import com.github.simbir_hospital_service.dto.HospitalDto;
-//import com.github.simbir_hospital_service.service.HospitalService;
-//import io.swagger.v3.oas.annotations.Operation;
-//import lombok.RequiredArgsConstructor;
-//import org.springframework.http.HttpStatus;
-//import org.springframework.http.ResponseEntity;
-//import org.springframework.web.bind.annotation.*;
-//
-//import java.util.List;
-//
-//@RestController
-//@RequiredArgsConstructor
-//@RequestMapping("/api/Hospitals")
-//public class HospitalController {
-//
-//    private final HospitalService hospitalService;
-//    private final AccountServiceClient accountServiceClient;
-//    private final UserContext userContext;
-//
-//    private boolean isUserAuthorized() {
-//        String token = userContext.getToken();
-//        if (token == null) {
-//            return false;
-//        }
-//
-//        ResponseEntity<?> validationResult = accountServiceClient.validateToken(token);
-//        return validationResult.getStatusCode().is2xxSuccessful();
-//    }
-//
-//    @Operation(summary = "Получение списка больниц")
-//    @GetMapping
-//    public ResponseEntity<List<Hospital>> getHospitals(
-//            @RequestParam int from,
-//            @RequestParam int count
-//    ) {
-//        if (!isUserAuthorized()) {
-//            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-//        }
-//        GetHospitalRequest getHospitalRequest = GetHospitalRequest
-//                .builder()
-//                .from(from)
-//                .count(count)
-//                .build();
-//
-//        List<Hospital> hospitals = hospitalService.getHospitals(getHospitalRequest);
-//        return ResponseEntity.ok(hospitals);
-//    }
-//
-//    @Operation(summary = "Получение информации о больнице по Id")
-//    @GetMapping("/{id}")
-//    public ResponseEntity<Hospital> getHospitalById(
-//            @PathVariable Long id) {
-//        Hospital hospital = hospitalService.getHospitalById(id);
-//        return ResponseEntity.ok(hospital);
-//    }
-//
-//    @Operation(summary = "Получение списка кабинетов больницы по Id")
-//    @GetMapping("/{id}/rooms")
-//    public ResponseEntity<List<String>> getHospitalRooms(@PathVariable Long id) {
-//        List<String> rooms = hospitalService.getHospitalRooms(id);
-//        return ResponseEntity.ok(rooms);
-//    }
-//
-//    @PostMapping
-//    public ResponseEntity<HospitalDto> createHospital(@RequestBody HospitalDto hospitalDto) {
-//        HospitalDto createdHospital = hospitalService.createHospital(hospitalDto);
-//        return ResponseEntity.status(HttpStatus.CREATED).body(createdHospital);
-//    }
-//
-//    @PutMapping("/{id}")
-//    public ResponseEntity<HospitalDto> updateHospital(@PathVariable Long id, @RequestBody HospitalDto hospitalDto) {
-//        return hospitalService.updateHospital(id, hospitalDto)
-//                .map(ResponseEntity::ok)
-//                .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
-//    }
-//
-//    @DeleteMapping("/{id}")
-//    public ResponseEntity<Void> deleteHospital(@PathVariable Long id) {
-//        hospitalService.deleteHospital(id);
-//        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-//    }
-//}
-
 package com.github.simbir_hospital_service.controller;
 
 import com.github.simbir_hospital_service.client.AccountServiceClient;
@@ -100,8 +11,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -112,21 +23,11 @@ import java.util.List;
 public class HospitalController {
 
     private final HospitalService hospitalService;
-    private final AccountServiceClient accountServiceClient;
     private final UserContext userContext;
 
-    private boolean isUserAuthorized() {
-        String token = userContext.getToken();
-        if (token == null) {
-            log.warn("Authorization token is missing.");
-            return false;
-        }
-
-        log.info("Sending token to AccountServiceClient for validation: {}", token);
-        ResponseEntity<?> validationResult = accountServiceClient.validateToken(token);
-        log.info("Token validation result status: {}", validationResult.getStatusCode());
-
-        return validationResult.getStatusCode().is2xxSuccessful();
+    private boolean isUserAuthorized(String role) {
+        List<String> roles = userContext.getRoles();
+        return roles.contains(role);
     }
 
     @Operation(summary = "Получение списка больниц")
@@ -152,6 +53,10 @@ public class HospitalController {
     @GetMapping("/{id}")
     public ResponseEntity<Hospital> getHospitalById(
             @PathVariable Long id) {
+
+        List<String> roles = userContext.getRoles();
+        log.info("Role:{} ", roles.get(0));
+
         log.info("Fetching hospital details for id: {}", id);
 
         Hospital hospital = hospitalService.getHospitalById(id);
@@ -172,8 +77,10 @@ public class HospitalController {
     }
 
     @PostMapping
-    @PreAuthorize("hasRole('admin')")
     public ResponseEntity<HospitalDto> createHospital(@RequestBody HospitalDto hospitalDto) {
+        if (!isUserAuthorized("admin")) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Доступ запрещен: недостаточно прав.");
+        }
         log.info("Creating a new hospital with name: {}", hospitalDto.name());
 
         HospitalDto createdHospital = hospitalService.createHospital(hospitalDto);
@@ -183,8 +90,10 @@ public class HospitalController {
     }
 
     @PutMapping("/{id}")
-    @PreAuthorize("hasRole('admin')")
     public ResponseEntity<HospitalDto> updateHospital(@PathVariable Long id, @RequestBody HospitalDto hospitalDto) {
+        if (!isUserAuthorized("admin")) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Доступ запрещен: недостаточно прав.");
+        }
         log.info("Updating hospital with id: {}", id);
 
         return hospitalService.updateHospital(id, hospitalDto)
@@ -199,8 +108,10 @@ public class HospitalController {
     }
 
     @DeleteMapping("/{id}")
-    @PreAuthorize("hasRole('admin')")
     public ResponseEntity<Void> deleteHospital(@PathVariable Long id) {
+        if (!isUserAuthorized("admin")) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Доступ запрещен: недостаточно прав.");
+        }
         log.info("Deleting hospital with id: {}", id);
 
         hospitalService.deleteHospital(id);
